// packages/form/src/pkg/types/api_types.ts
export interface FileMetadata {
	eTag: string;
	size: bigint;
	mimetype: string;
	cacheControl: string;
	lastModified?: Timestamp;
	contentLength: bigint;
	httpStatusCode: number;
}

export type StorageBucket = {
	id: string;
	name: string;
	public: boolean;
	createdAt: string;
	updatedAt: string;
	owner: string;
	fileSizeLimit: bigint;
	allowedMimeTypes: string[];
}
export type FileObject = {
	name: string;
	bucketId: string;
	owner: string;
	id: string;
	updatedAt?: string;
	createdAt?: string;
	lastAccessedAt?: string;
	metadata?: FileMetadata; // Assuming FileMetadata is another defined type
	buckets?: StorageBucket; // Assuming StorageBucket is another defined type
}
export type FileListResponse<T extends string = 'records'> = {
	[K in T]: FileObject[]
} & {
	options?: StringUnknownRecord
}
export type FileCreateRequest = {
	path: string;
	bucketName: string;
	reader: Uint8Array;
	fileType: string;
}
export type FileCreateResponse = {
	path: string;
}
export type FileCreateBulkRequest = {
	files: FileCreateRequest[];
}